# server/docs/swagger.yaml
basePath: /
definitions:
  schema.ChartData:
    properties:
      label:
        description: day
        type: string
      value:
        description: counter in/out
        type: integer
    type: object
  schema.ChartDataValue:
    properties:
      in:
        $ref: '#/definitions/schema.ChartData'
      out:
        $ref: '#/definitions/schema.ChartData'
    type: object
  schema.ChartMetadata:
    properties:
      data:
        items:
          $ref: '#/definitions/schema.ChartDataValue'
        type: array
      month:
        type: string
      room_id:
        type: string
      room_name:
        type: string
      year:
        type: string
    type: object
  schema.CheckInRequest:
    properties:
      activity_type:
        type: string
      admin_id:
        type: string
      registration_id:
        type: string
    required:
    - activity_type
    - admin_id
    - registration_id
    type: object
  schema.DataResponse:
    properties:
      data: {}
      success:
        type: boolean
    required:
    - data
    - success
    type: object
  schema.HistoryData:
    properties:
      activity_type:
        type: string
      admin_id:
        type: string
      email:
        type: string
      end_day:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      office_name:
        type: string
      org_name:
        type: string
      phone_number:
        type: string
      registration_time:
        type: string
      room_name:
        type: string
      start_day:
        type: string
      supervisor:
        type: string
      time:
        type: string
      user_id:
        type: string
    type: object
  schema.HistoryDataResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/schema.HistoryData'
        type: array
      room_name:
        type: string
      total_in:
        type: integer
      total_out:
        type: integer
    type: object
  schema.RegistrationRoomRequest:
    properties:
      email:
        type: string
      end_day:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      phone_number:
        type: string
      registration_time:
        type: string
      room_id:
        type: string
      start_day:
        type: string
      student_id:
        type: string
      supervisor:
        type: string
    required:
    - email
    - end_day
    - first_name
    - last_name
    - phone_number
    - registration_time
    - room_id
    - start_day
    - student_id
    - supervisor
    type: object
  schema.Response:
    properties:
      msg:
        type: string
      success:
        type: boolean
    required:
    - msg
    - success
    type: object
  schema.RoomStat:
    properties:
      data:
        items:
          $ref: '#/definitions/schema.RoomStatData'
        type: array
      total:
        type: integer
    type: object
  schema.RoomStatData:
    properties:
      in:
        type: integer
      out:
        type: integer
      room_id:
        type: string
      room_name:
        type: string
      total:
        type: integer
    type: object
  schema.SignInRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  schema.SignInResponse:
    properties:
      email:
        type: string
      success:
        type: boolean
      token:
        type: string
    required:
    - email
    - success
    - token
    type: object
  schema.SignUpRequest:
    properties:
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
    required:
    - email
    - first_name
    - id
    - last_name
    - password
    - phone_number
    type: object
  schema.SignUpResponse:
    properties:
      msg:
        type: string
      success:
        type: boolean
    required:
    - msg
    - success
    type: object
info:
  contact: {}
  description: This is a documentation for the Student Checkin System API
  title: Student Checkin System
  version: 1.0.0
paths:
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Sign in account for admin.
      parameters:
      - description: Sign In
        in: body
        name: sign_in
        required: true
        schema:
          $ref: '#/definitions/schema.SignInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.SignInResponse'
      tags:
      - auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Register account for admin.
      parameters:
      - description: Sign Up
        in: body
        name: sign_up
        required: true
        schema:
          $ref: '#/definitions/schema.SignUpRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.SignUpResponse'
      tags:
      - auth
  /common/healthcheck:
    get:
      consumes:
      - application/json
      description: health check api server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
      tags:
      - common
  /room:
    get:
      consumes:
      - application/json
      description: Get Room information from database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.DataResponse'
      tags:
      - room
  /room/activity:
    get:
      consumes:
      - application/json
      description: Get User Activity for checkin.
      parameters:
      - description: student id
        in: query
        name: sid
        required: true
        type: string
      - description: room id
        in: query
        name: room
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.DataResponse'
      tags:
      - room
    post:
      consumes:
      - application/json
      description: Save activity type in/out.
      parameters:
      - description: CheckInRequest
        in: body
        name: activity_param
        required: true
        schema:
          $ref: '#/definitions/schema.CheckInRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.Response'
      tags:
      - room
  /room/history:
    get:
      consumes:
      - application/json
      description: Get History.
      parameters:
      - description: room id
        in: query
        name: room_id
        required: true
        type: string
      - description: date only
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.HistoryDataResponse'
      tags:
      - room
  /room/register:
    post:
      consumes:
      - application/json
      description: Register Room.
      parameters:
      - description: RegisterRoom
        in: body
        name: sign_in
        required: true
        schema:
          $ref: '#/definitions/schema.RegistrationRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.Response'
      tags:
      - room
  /stat/chart:
    get:
      consumes:
      - application/json
      description: Get Chart Data IN/OUT.
      parameters:
      - description: room id
        in: query
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.ChartMetadata'
      tags:
      - stat
  /stat/rooms:
    get:
      consumes:
      - application/json
      description: Get Room Data IN/OUT.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.RoomStat'
      tags:
      - stat
  /user:
    get:
      consumes:
      - application/json
      description: Get information by student id
      parameters:
      - description: student id
        in: query
        name: sid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schema.SignInResponse'
      tags:
      - user
swagger: "2.0"